- Added utility function to Resources class. - Factory Production window will only show if the factory is right clicked on and it's either operational or idle. - Factory is now documented much better especially the assumptions that it makes. - Added a number of utility and get/set type functions to Factory. - Factory will now fill a production 'table' with production costs and times for various productions. Currently only defines the Digger/Dozer/Miner robots. - Updated button so that it uses whole numbers when centering text so we can avoid the blurring that comes with floats. - FactoryProduction now overrides hide() function. Clean, easy way to clear some values we don't need when hidden. - FactoryProduction now sets and gets production type from any factory object it's interacting with. - Added a set function to IconGrid to set the current selection. - Game now searches for the following data archives when starting up: "fonts.dat", "maps.dat", "planets.dat", "structures.dat", "robots.dat, "ui.dat". What format these archives are in is irrelevant so long as they are archive formats supported by the PhysicsFS filesystem library. If these archives aren't found OutpostHD will continue to look in the normal filesystem for these files.
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@36 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/UI/Core/Button.cpp
@@ -150,7 +150,7 @@ void Button::draw()
 		r.drawImage(mImage, rect().x() + (rect().w() / 2) - (mImage.width() / 2), rect().y() + (rect().h() / 2) - (mImage.height() / 2));
 	else
 		if (fontSet() && !text().empty())
-			r.drawText(font(), text(), rect().x() + (rect().w() / 2) - (font().width(text()) / 2), rect().y() + (rect().h() / 2) - (font().height() / 2), 0, 0, 0);
+			r.drawText(font(), text(), static_cast<int>(rect().x() + (rect().w() / 2) - (font().width(text()) / 2)), static_cast<int>(rect().y() + (rect().h() / 2) - (font().height() / 2)), 0, 0, 0);
 
 	if (!enabled())
 		r.drawBoxFilled(rect(), 200, 200, 200, 100);
